Use redux toolkit for global store
