--- src/store/GlobalStore/GlobalHooks.ts
@@ -1,44 +1,0 @@
-import { useCallback } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import {
-  acceptCookieConcentAction,
-  closeNavigationAction,
-  toggleNavigationAction,
-} from "./GlobalActions";
-import {
-  acceptedCookieConsentSelector,
-  isNavOpenSelector,
-} from "./GlobalSelectors";
-
-export const useNavigation = () => {
-  const dispatch = useDispatch();
-  const isNavOpen = useSelector(isNavOpenSelector);
-
-  const toggleNav = useCallback(() => {
-    dispatch(toggleNavigationAction());
-  }, [dispatch]);
-
-  const closeNav = useCallback(() => {
-    dispatch(closeNavigationAction());
-  }, [dispatch]);
-
-  return {
-    isNavOpen,
-    toggleNav,
-    closeNav,
-  };
-};
-
-export const useCookieConcent = () => {
-  const dispatch = useDispatch();
-  const acceptedCookieConsent = useSelector(acceptedCookieConsentSelector);
-
-  const acceptCookieConcent = useCallback(() => {
-    dispatch(acceptCookieConcentAction());
-  }, [dispatch]);
-
-  return {
-    acceptedCookieConsent,
-    acceptCookieConcent,
-  };
-};
