--- src/store/GlobalStore/hooks/helpers.ts
@@ -1,6 +1,0 @@
-import { useDispatch } from "react-redux";
-import { Dispatch } from "redux";
-
-export const useGlobalDispatch = () => {
-  return useDispatch<Dispatch<any>>();
-};
