+++ src/store/rootReducer.ts
@@ -1,12 +1,12 @@
 import { combineReducers } from "redux";
 import { FavoriteReducer } from "./FavoriteStore/FavoriteReducer";
-import { GlobalReducer } from "./GlobalStore/GlobalReducer";
 import { NotificationReducer } from "./NotificationStore/NotificationReducer";
 import { ToastReducer } from "./ToastStore/ToastReducer";
 import { MapStoreReducer } from "./MapStore/slice";
 import TrackerReducer from "../features/tracker/TrackerSlice";
 import { SettingsReducer } from "./SettingsStore/SettingsSlice";
 import { AuthReducer } from "./AuthStore/AuthSlice";
+import { GlobalReducer } from "./GlobalStore/GlobalStore";
 
 const rootReducer = combineReducers({
   authStore: AuthReducer,
